+++ packages/jsenv-plugin-toolbar/src/client/core/animation_actions.js
@@ -1,0 +1,8 @@
+import { animationsEnabledSignal } from "./animation_signals.js"
+
+export const enableAnimations = () => {
+  animationsEnabledSignal.value = true
+}
+export const disableAnimations = () => {
+  animationsEnabledSignal.value = false
+}
