+++ packages/jsenv-plugin-toolbar/src/client/core/animation_signals.js
@@ -1,0 +1,3 @@
+import { signal } from "@preact/signals"
+
+export const animationsEnabledSignal = signal(false)
