+++ packages/jsenv-plugin-toolbar/src/client/core/notification_context.js
@@ -1,0 +1,1 @@
+export const notificationAPIDetected = typeof window.Notification === "function"
