+++ packages/jsenv-plugin-toolbar/src/client/core/notification_signals.js
@@ -1,0 +1,4 @@
+import { signal } from "@preact/signals"
+
+export const notificationsEnabledSignal = signal(false)
+export const notificationPermissionSignal = signal(Notification.permission)
