+++ packages/jsenv-plugin-toolbar/src/client/core/theme_actions.js
@@ -1,0 +1,9 @@
+import { themeSignal } from "./theme_signals.js"
+
+export const switchToLightTheme = () => {
+  themeSignal.value = "light"
+}
+
+export const switchToDefaultTheme = () => {
+  themeSignal.value = "dark"
+}
