+++ packages/jsenv-plugin-toolbar/src/client/core/theme_signals.js
@@ -1,0 +1,3 @@
+import { signal } from "@preact/signals"
+
+export const themeSignal = signal("dark")
