+++ packages/jsenv-plugin-toolbar/src/client/core/toolbar_open_actions.js
@@ -1,0 +1,9 @@
+import { openedSignal } from "./toolbar_open_signals.js"
+
+export const openToolbar = () => {
+  openedSignal.value = true
+}
+
+export const closeToolbar = () => {
+  openedSignal.value = false
+}
