+++ packages/jsenv-plugin-toolbar/src/client/core/toolbar_open_effects.js
@@ -1,0 +1,10 @@
+import { stateFromLocalStorage } from "./toolbar_state_context.js"
+import { openedSignal } from "./toolbar_open_signals.js"
+import { addExternalCommandCallback } from "./parent_window_communication.js"
+import { openToolbar, closeToolbar } from "./toolbar_open_actions.js"
+
+if (stateFromLocalStorage.opened) {
+  openedSignal.value = true
+}
+addExternalCommandCallback("openToolbar", openToolbar)
+addExternalCommandCallback("closeToolbar", closeToolbar)
