+++ packages/jsenv-plugin-toolbar/src/client/core/toolbar_open_signals.js
@@ -1,0 +1,3 @@
+import { signal } from "@preact/signals"
+
+export const openedSignal = signal(false)
