+++ packages/jsenv-plugin-toolbar/src/client/core/toolbar_state_context.js
@@ -1,0 +1,5 @@
+export const stateFromLocalStorage = localStorage.hasOwnProperty(
+  "jsenv_toolbar",
+)
+  ? JSON.parse(localStorage.getItem("jsenv_toolbar"))
+  : {}
